BREAKING(flutter upgrade 19.1): flutter sdk updated to 19.1 from 13. . . . New ServerService to centralize http requests

+++ market_check/lib/features/profile/data/datasources/profile_data_source.dart
@@ -1,37 +1,34 @@
 import 'dart:convert';
-
 import 'package:flutter/foundation.dart';
 
 import 'package:market_check/config/errors/exceptions.dart';
 import 'package:market_check/config/shared/models/user.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:market_check/config/services/auth/auth_service.dart';
-import 'package:market_check/config/services/remote_service/remote_urls.dart';
+import 'package:market_check/config/services/server/server_urls.dart';
+import 'package:market_check/config/services/server/server_service.dart';
 import 'package:market_check/config/shared/models/create_user_data_model.dart';
-
-import 'package:dio/dio.dart';
-import 'package:http/http.dart' as http;
 
 abstract class ProfileDataSource {
   Future<bool> updatePasword(String password);
-  Future<String> deleteAccount();
   Future<String> updateAccountData(SignUpDataModel updatedData);
+  Future<String> deleteAccount();
 }
 
 class ProfileDataSourceImpl extends ProfileDataSource {
   final FlutterSecureStorage flutterSecureStorage;
 
-  final Dio dioInstance;
-
-  ProfileDataSourceImpl(
-      {required this.flutterSecureStorage, required this.dioInstance});
+  ProfileDataSourceImpl({required this.flutterSecureStorage});
 
   @override
   Future<bool> updatePasword(String password) async {
     try {
-      Response response = await dioInstance.put(
-        '${RemoteUrls.userUrl}${AuthService.user!.id!}/change-password',
-        data: {'password': password},
+      final String path =
+          '${ServerUrls.userUrl}${AuthService.user!.id!}/change-password';
+
+      final response = await ServerService.serverPut(
+        path,
+        {'password': password},
       );
 
       if (response.statusCode == 200 || response.statusCode == 201) {
@@ -39,8 +36,14 @@ class ProfileDataSourceImpl extends ProfileDataSource {
       }
 
       return false;
+    } on HttpException catch (e) {
+      debugPrint('$e');
+      throw RemoteException(
+          message:
+              "Ocurrio un error al conectarse al servidor, intente de nuevo mas tarde",
+          type: ExceptionType.signInException);
     } catch (e) {
-      print(e);
+      debugPrint('$e');
       throw RemoteException(
           message: "Ocurrio un error al actualizar la contraseña",
           type: ExceptionType.signInException);
@@ -50,20 +53,11 @@ class ProfileDataSourceImpl extends ProfileDataSource {
   @override
   Future<String> updateAccountData(SignUpDataModel updatedData) async {
     try {
-      /*final Response response = await dioInstance.put(
-        '${RemoteUrls.userUrl}${AuthService.user!.id!}',
-        data: updatedData.userToJson(),
-      );*/
-      //TODO TERMINAR
-      var url = Uri.http(
-          RemoteUrls.currentHttp, '/api/${RemoteUrls.userUrl}${AuthService.user!.id!}');
-      
-      var response = await http.put(
-        url,
-        headers: AuthService.headers,
-        body: jsonEncode(
-          updatedData.userToJson(),
-        ),
+      final String path = '${ServerUrls.userUrl}${AuthService.user!.id!}';
+
+      var response = await ServerService.serverPut(
+        path,
+        updatedData.userToJson(),
       );
 
       if (response.statusCode == 201) {
@@ -78,14 +72,14 @@ class ProfileDataSourceImpl extends ProfileDataSource {
             key: 'email', value: updatedUser.email);
       }
       return jsonDecode(response.body)["message"];
-    } on DioException catch (e) {
+    } on HttpException catch (e) {
       debugPrint('$e');
       throw RemoteException(
           message:
-              "Ocurrio un error al actualizar los datos, por favor intente de nuevo",
+              "Ocurrio un error al conectarse al servidor, intente de nuevo mas tarde",
           type: ExceptionType.signInException);
     } catch (e) {
-      print(e);
+      debugPrint('$e');
       throw RemoteException(
           message: "Ocurrio un error al actualizar los datos",
           type: ExceptionType.signInException);
@@ -95,20 +89,23 @@ class ProfileDataSourceImpl extends ProfileDataSource {
   @override
   Future<String> deleteAccount() async {
     try {
-      final Response response = await dioInstance
-          .delete('${RemoteUrls.userUrl}${AuthService.user!.id}');
+      final String path = '${ServerUrls.userUrl}${AuthService.user!.id}';
 
-      await flutterSecureStorage.deleteAll();
+      final response = await ServerService.serverDelete(path);
 
-      return response.data["message"];
-    } on DioException catch (e) {
-      print(e);
+      if (response.statusCode == 200 || response.statusCode == 201) {
+        await flutterSecureStorage.deleteAll();
+      }
+
+      return jsonDecode(response.body)["message"];
+    } on HttpException catch (e) {
+      debugPrint('$e');
       throw RemoteException(
           message:
-              "Ocurrio un error al eliminar la cuenta, por favor intentalo despues.",
+              "Ocurrio un error al conectarse al servidor, intente de nuevo mas tarde",
           type: ExceptionType.signInException);
     } catch (e) {
-      print(e);
+      debugPrint('$e');
       throw RemoteException(
           message:
               "Ocurrio un error al eliminar la cuenta, por favor intentalo despues.",
